+++ plato/src/main/java/jstorra/plato/repositories/PlatoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlatoRepository extends JpaRepository<Plato, Long> {
+public interface PlatoRepository extends JpaRepository<Plato, Integer> {
 }
